+++ src/js/components/overview/OverviewDisplayData.js
@@ -2,120 +2,17 @@ import React from 'react';
 import BentoBarChart from './charts/BentoBarChart';
 import BentoPie from './charts/BentoPie';
 import { Col } from 'react-bootstrap';
+import {
+  orderCategories,
+  parseDateData,
+  parseNumericalData,
+  parseDataGeneral,
+  binMissingData,
+} from '../../utils/DataUtils';
 
 const CHART_HEIGHT = 300;
 
 const OverviewDisplayData = ({ all_charts, queryParameterStack }) => {
-  const orderCategories = (data, ordering) =>
-    data.sort((a, b) => {
-      let aIndex = ordering.indexOf(a.x);
-      let bIndex = ordering.indexOf(b.x);
-
-      // send any categories not found to end of array
-      if (aIndex < 0) {
-        aIndex = Number.MAX_VALUE;
-      }
-      if (bIndex < 0) [(bIndex = Number.MAX_VALUE)];
-      return aIndex - bIndex;
-    });
-
-  const parseDateData = (inputData) => {
-    const formattedData = [];
-
-    Object.keys(inputData).forEach((_key) => {
-      const attemptedDate = new Date(_key);
-      const monthStr = attemptedDate.yyyydashmm();
-
-      // if not a valid date, add to "missing"
-      if (isNaN(attemptedDate.valueOf())) {
-        monthlybins.missing = (monthlybins.missing ?? 0) + inputData[_key];
-        return; // ie, continue
-      }
-
-      // verify if monthly bin exists - add to dict if not
-      if (monthlybins[monthStr] == undefined) {
-        monthlybins[monthStr] = inputData[_key];
-      } else {
-        monthlybins[monthStr] = monthlybins[monthStr] + inputData[_key];
-      }
-    });
-
-    Object.keys(monthlybins).forEach((_key) => {
-      formattedData.push({ x: _key, y: monthlybins[_key] });
-    });
-
-    return formattedData;
-  };
-
-  const parseDataGeneral = (
-    inputData,
-    leftTaper,
-    rightTaper,
-    binSize,
-    field
-  ) => {
-    let formattedData = [];
-    let taperLeft, taperRight;
-
-    Object.keys(inputData).forEach((key, index, arr) => {
-      const intKey = parseInt(key);
-      if (intKey != NaN && field.type == 'number') {
-        if (intKey < leftTaper) {
-          const tlkey = `< ${leftTaper}`;
-          if (taperLeft == undefined) {
-            taperLeft = { x: tlkey, y: inputData[key] };
-          } else {
-            taperLeft.y += inputData[key];
-          }
-        } else if (intKey >= rightTaper + binSize) {
-          const trkey = `>= ${rightTaper + binSize}`;
-          if (taperRight == undefined) {
-            taperRight = { x: trkey, y: inputData[key] };
-          } else {
-            taperRight.y += inputData[key];
-          }
-        } else {
-          // number range tag
-          let xTag = '';
-          const leftVal = key;
-          // add rightVal if exists
-          if (index + 1 < arr.length) {
-            const rightVal = arr[index + 1];
-            xTag = `${leftVal} - ${rightVal}`;
-          } else {
-            xTag = leftVal;
-          }
-
-          formattedData.push({ x: xTag, y: inputData[key] });
-        }
-      } else {
-        formattedData.push({ x: key, y: inputData[key] });
-      }
-    });
-
-    if (taperLeft != undefined) {
-      // prepend left taper
-      formattedData.unshift(taperLeft);
-    }
-
-    if (taperRight != undefined) {
-      // append right taper
-      formattedData.push(taperRight);
-    }
-
-    return formattedData;
-  };
-
-  const binMissingData = (data) => {
-    const missingObject = data.filter((obj) => obj.x === 'missing');
-
-    if (missingObject.length > 0) {
-      // fancy one liner to splice the object out and re-append it
-      data.push(data.splice(data.indexOf(missingObject[0]), 1)[0]);
-    }
-    return data;
-  };
-
   return (
     <>
       {
@@ -149,14 +46,15 @@ const OverviewDisplayData = ({ all_charts, queryParameterStack }) => {
 
             if (field.type == 'string' && field.format == 'date') {
               formattedData = parseDateData(inputData);
-            } else {
-              formattedData = parseDataGeneral(
+            } else if (field.type == 'numerical') {
+              formattedData = parseNumericalData(
                 inputData,
                 leftTaper,
                 rightTaper,
-                binSize,
-                field
+                binSize
               );
+            } else {
+              formattedData = parseDataGeneral(inputData);
             }
 
             // move 'missing' object to end of list
