+++ src/js/components/overview/PublicOverview.js
@@ -1,20 +1,32 @@
-// Dashboard.js
 import React from 'react';
 import { Row, Container } from 'react-bootstrap';
 import { Divider, Typography } from 'antd';
 
+import { parseData } from '../../utils/DataUtils';
+
 import OverviewDisplayData from './OverviewDisplayData';
 
 const PublicOverview = ({ overview, queryParameterStack }) => {
-  const ep = overview.extra_properties;
-  const all_charts = { ...overview, ...ep };
-  const something = [
-    { name: 'Age', data: overview.age },
-    { name: 'Sex', data: overview.sex },
-    ...Object.entries(ep).map((item) => ({ name: item[0], data: item[1] })),
-  ];
+  const all_charts_obj = { ...overview, ...overview.extra_properties };
+  delete all_charts_obj.extra_properties;
+  delete all_charts_obj.individuals;
 
-  console.log('something', something);
+  const all_charts = Object.entries(all_charts_obj).map((item) => ({
+    name: item[0],
+    data: item[1],
+  }));
+
+  all_charts.forEach((chart, index, arr) => {
+    arr[index].properties = queryParameterStack.find(
+      (e) => e?.key == chart.name
+    )?.props?.Item;
+  });
+
+  all_charts.forEach((chart, index, arr) => {
+    arr[index].data = parseData(chart);
+  });
+
+  console.log('data', all_charts);
 
   return (
     <Container>
@@ -31,7 +43,7 @@ const PublicOverview = ({ overview, queryParameterStack }) => {
           </Typography.Title>
         </div>
         <OverviewDisplayData
-          all_charts={all_charts}
+          all_charts={all_charts_obj}
           queryParameterStack={queryParameterStack}
         />
       </Row>
